+++ src/tkc/named_value_hash.c
@@ -24,21 +24,25 @@
 
 #include "tkc/named_value_hash.h"
 
-uint64_t named_value_hash_get_hash_from_str(const char* str, uint64_t base) {
-  uint64_t ret = 0;
-  uint64_t p = 0;
+static uint64_t bkdr_hash(const char* str) {
+  const uint64_t base = 131;
   const char* iter = NULL;
-  for (iter = str, p = 1; (*iter) != '\0'; iter++, p *= base) {
-    ret += (*iter) * p;
+  uint64_t ret = 0;
+  for (ret = 0, iter = str; (*iter) != '\0'; iter++) {
+    ret = ret * base + (*iter);
   }
   return ret;
 }
 
-ret_t named_value_hash_set_name(named_value_hash_t* nvh, const char* name, uint64_t hash_base) {
+uint64_t named_value_hash_get_hash_from_str(const char* str) {
+  return bkdr_hash(str);
+}
+
+ret_t named_value_hash_set_name(named_value_hash_t* nvh, const char* name) {
   return_value_if_fail(nvh != NULL, RET_BAD_PARAMS);
 
   named_value_set_name(&nvh->base, name);
-  nvh->hash = named_value_hash_get_hash_from_str(name, hash_base);
+  nvh->hash = named_value_hash_get_hash_from_str(name);
 
   return RET_OK;
 }
@@ -53,14 +57,14 @@ ret_t named_value_hash_deinit(named_value_hash_t* nvh) {
 }
 
 named_value_hash_t* named_value_hash_init(named_value_hash_t* nvh, const char* name,
-                                          const value_t* value, uint64_t hash_base) {
+                                          const value_t* value) {
   ret_t ret = RET_OK;
   return_value_if_fail(nvh != NULL, NULL);
 
   memset(nvh, 0x00, sizeof(named_value_hash_t));
 
   if (name != NULL) {
-    ret = named_value_hash_set_name(nvh, name, hash_base);
+    ret = named_value_hash_set_name(nvh, name);
   }
 
   if (RET_OK == ret) {
@@ -81,12 +85,11 @@ ret_t named_value_hash_destroy(named_value_hash_t* nvh) {
   return RET_OK;
 }
 
-named_value_hash_t* named_value_hash_create_ex(const char* name, const value_t* value,
-                                               uint64_t hash_base) {
+named_value_hash_t* named_value_hash_create_ex(const char* name, const value_t* value) {
   named_value_hash_t* ret = TKMEM_ZALLOC(named_value_hash_t);
   return_value_if_fail(ret != NULL, NULL);
 
-  goto_error_if_fail(ret == named_value_hash_init(ret, name, value, hash_base));
+  goto_error_if_fail(ret == named_value_hash_init(ret, name, value));
 
   return ret;
 error:
@@ -95,19 +98,34 @@ error:
 }
 
 named_value_hash_t* named_value_hash_create(void) {
-  return named_value_hash_create_ex(NULL, NULL, HASH_BASE_DEFAULT);
+  return named_value_hash_create_ex(NULL, NULL);
 }
 
-int32_t named_value_hash_compare(named_value_hash_t* nvh, const named_value_hash_t* other) {
+static int32_t named_value_hash_hash_compare(uint64_t hash1, uint64_t hash2) {
+  if (hash1 > hash2) {
+    // uint64_t d = hash1 - hash2;
+    // return tk_min(d, INT32_MAX);
+    return 1;
+  } else if (hash1 < hash2) {
+    // uint64_t d = hash2 - hash1;
+    // int64_t t = -tk_min(d, -(int64_t)INT32_MIN);
+    // return tk_max(INT32_MIN, t);
+    return -1;
+  } else {
+    return 0;
+  }
+}
+
+int32_t named_value_hash_compare(const named_value_hash_t* nvh, const named_value_hash_t* other) {
   return_value_if_fail(nvh != NULL && other != NULL, -1);
 
-  return nvh->hash - other->hash;
+  return named_value_hash_hash_compare(nvh->hash, other->hash);
 }
 
-int32_t named_value_hash_compare_by_hash(named_value_hash_t* nvh, uint64_t hash) {
+int32_t named_value_hash_compare_by_hash(const named_value_hash_t* nvh, uint64_t hash) {
   return_value_if_fail(nvh != NULL, -1);
 
-  return nvh->hash - hash;
+  return named_value_hash_hash_compare(nvh->hash, hash);
 }
 
 named_value_hash_t* named_value_hash_clone(named_value_hash_t* nvh) {
@@ -115,9 +133,7 @@ named_value_hash_t* named_value_hash_clone(named_value_hash_t* nvh) {
   return_value_if_fail(ret != NULL, NULL);
 
   ret->hash = nvh->hash;
-
   named_value_set_name(&ret->base, nvh->base.name);
-
   named_value_set_value(&ret->base, &nvh->base.value);
 
   return ret;