+++ tests/named_value_hash_test.cc
@@ -9,7 +9,7 @@ TEST(NamedValueHash, basic) {
   named_value_hash_t named_value_hash;
   named_value_hash_t* nvh = &named_value_hash;
 
-  named_value_hash_init(nvh, NULL, NULL, HASH_BASE_DEFAULT);
+  named_value_hash_init(nvh, NULL, NULL);
 
   ASSERT_EQ(nvh->hash, 0);
   ASSERT_EQ(nvh->base.name, (char*)NULL);
@@ -24,9 +24,9 @@ TEST(NamedValueHash, init) {
   named_value_hash_t* nvh = &named_value_hash;
 
   value_set_int(&v, 123);
-  named_value_hash_init(nvh, "name", &v, HASH_BASE_DEFAULT);
+  named_value_hash_init(nvh, "name", &v);
 
-  ASSERT_EQ(nvh->hash, named_value_hash_get_hash_from_str("name", HASH_BASE_DEFAULT));
+  ASSERT_EQ(nvh->hash, named_value_hash_get_hash_from_str("name"));
   ASSERT_EQ(string(nvh->base.name), string("name"));
   ASSERT_EQ(value_int(&v), value_int(&(nvh->base.value)));
 
@@ -39,10 +39,9 @@ TEST(NamedValueHash, init) {
 
 TEST(NamedValueHash, create_ex) {
   value_t v;
-  named_value_hash_t* nvh =
-      named_value_hash_create_ex("name", value_set_str(&v, "jim"), HASH_BASE_DEFAULT);
+  named_value_hash_t* nvh = named_value_hash_create_ex("name", value_set_str(&v, "jim"));
 
-  ASSERT_EQ(nvh->hash, named_value_hash_get_hash_from_str("name", HASH_BASE_DEFAULT));
+  ASSERT_EQ(nvh->hash, named_value_hash_get_hash_from_str("name"));
   ASSERT_STREQ(nvh->base.name, "name");
   ASSERT_STREQ(value_str(&(nvh->base.value)), "jim");
 
@@ -51,14 +50,13 @@ TEST(NamedValueHash, create_ex) {
 
 TEST(NamedValueHash, create_ex1) {
   value_t v;
-  named_value_hash_t* nvh =
-      named_value_hash_create_ex("name", value_set_str(&v, "jim"), HASH_BASE_DEFAULT);
+  named_value_hash_t* nvh = named_value_hash_create_ex("name", value_set_str(&v, "jim"));
 
-  ASSERT_EQ(nvh->hash, named_value_hash_get_hash_from_str("name", HASH_BASE_DEFAULT));
+  ASSERT_EQ(nvh->hash, named_value_hash_get_hash_from_str("name"));
   ASSERT_STREQ(nvh->base.name, "name");
   ASSERT_STREQ(value_str(&(nvh->base.value)), "jim");
-  ASSERT_EQ(named_value_hash_set_name(nvh, "tom", HASH_BASE_DEFAULT), RET_OK);
-  ASSERT_EQ(named_value_hash_set_name(nvh, "jike", HASH_BASE_DEFAULT), RET_OK);
+  ASSERT_EQ(named_value_hash_set_name(nvh, "tom"), RET_OK);
+  ASSERT_EQ(named_value_hash_set_name(nvh, "jike"), RET_OK);
 
   named_value_hash_destroy(nvh);
 }